New shot at the actor model.
The problem with the current implementation is that all the work is done
by the Context (now called Dataset :wink:) Actor and thus all the
webservice has become mainly single threaded...

This PR introduces:
 * a DatasetActor whose only job is to give an Arc (a shared pointer) to
a Dataset
 * a BaseScheduleReloader Actor whose job is to periodically reload the
data, and send the newly created Dataset to the DatasetActor for
updating.

Since the DatasetActor works in its own thread, this way we'll be able
to remove all the Mutex, and still have a multithread service!

For the moment all the Arcs have not been removed, because we still need
the mutate the Dataset when receiving a gtfs-rt, but this will be
changed in another PR.