multidatasets
enable the API to load multiple datasets

⚠️ this is a breaking API change ⚠️

now the Apis are nested in a `/datasets/{id}` path.

example:
`/datasets/metromobilite/siri-lite/stoppoints_discovery.json?q=mairie`

it also adds  `/datasets` route to have the list of the loaded datasets

The datasets are given in a yaml file. Our configuration is store in the
depository for ease of use.

Note: we have to update clever cloud after merging this PR