add some basic sentry
sentry can be activated with the --sentry parameter (or
TRANSPO_RT_SENTRY env var)

if activated sentry will now report:

* the 5xx errors
* the gtfs_rt reloading errors
* a warning if some stop_id are incoherent

for the moment for the gtfs_rt problems, we add the dataset id as a tag
to the error. we'll need to see if we need more information

some more check will need to be added in the future.