handle errors while reloading base schedule
the base schedules are reloading once per day. This operation can fail,
if for example the external  erver is temporarily unavailable.

If the reloading fails, we try it again 5mn later, while just keeping
the old dataset.

It is not exponential back-off nor do we particularly monitor those
fails. If they become more troublesome we could implement more advanced
techniques.