Fix url_for bug
linked to https://github.com/actix/actix-web/issues/1763

There was a bug, all links were linking to the same datasets (because
the routes are scoped).

I don't really see a clean way to make actix handle this, so for the
moment I use a custom name for each route based on the scope
(`{scope}/{route_name}`).

With this we can't use the niiiice `#[get("/routes")]` anymore :sob: